+++ attack/AttackBehavior.java
@@ -1,4 +1,6 @@
-package gameCharacters;
+package attack;
+
+import gameCharacters.DungeonCharacter;
 
 public interface AttackBehavior {
     int attack(DungeonCharacter enemy);
