+++ attack/Club.java
@@ -1,6 +1,9 @@
-package gameCharacters;
+package attack;
 
-public class LongSword implements AttackBehavior {
+import main.Dice;
+import gameCharacters.DungeonCharacter;
+
+public class Club implements AttackBehavior {
 
     public int attack(DungeonCharacter enemy) {
         boolean doesHit = Dice.d20() - enemy.getAC() > 0;
@@ -8,17 +11,18 @@ public class LongSword implements AttackBehavior {
             int dam = 2 + Dice.d8();
             enemy.subtractHitPoints(dam);
             return dam;
-        } else {
-            return 0;
         }
+        return 0;
     }
 
     @Override
-    public String toString() {
-        return " slashes ";
+    public String getName() {
+        return "Club";
     }
 
-    public String getName() {
-        return "Great sword";
+    public String toString() {
+        return " bashes ";
     }
 }
+
+
