+++ gameCharacters/Aarakocra.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Aarakocra extends Hero{
 	
 	public Aarakocra(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
