+++ gameCharacters/BugBear.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class BugBear extends Hero {
 
 	public BugBear(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
