+++ gameCharacters/Dwarf.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Dwarf extends Hero {
 
 	public Dwarf(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
