+++ gameCharacters/Elf.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Elf extends Hero {
 	
 	public Elf(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
