+++ gameCharacters/Gnome.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Gnome extends Hero {
 
 	public Gnome(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
