+++ gameCharacters/Goliath.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Goliath extends Hero {
 
 	public Goliath(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
