+++ gameCharacters/HalfElf.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class HalfElf extends Hero {
 
 	public HalfElf(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
