+++ gameCharacters/Human.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Human extends Hero {
 
 	public Human(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
