+++ gameCharacters/Lizardfolk.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Lizardfolk extends Hero {
 
 	public Lizardfolk(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
