+++ gameCharacters/Monster.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Monster extends DungeonCharacter {
 
 	 public Monster(String name, int hitPoints, int initiative, int ac, AttackBehavior ab) {
