+++ gameCharacters/Skeleton.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.LongSword;
+
 public class Skeleton extends Monster{
     
 	public Skeleton()
