+++ gameCharacters/Tabaxi.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Tabaxi extends Hero {
 
 	public Tabaxi(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
