+++ gameCharacters/Triton.java
@@ -1,5 +1,7 @@
 package gameCharacters;
 
+import attack.AttackBehavior;
+
 public class Triton extends Hero {
 	
 	public Triton(String name, int hitPoints, int initiative, int ac, String charClass, AttackBehavior ab) {
