+++ specialAttack/DaggerStorm.java
@@ -1,11 +1,14 @@
-package gameCharacters;
+package specialAttack;
+
+import main.Dice;
+import gameCharacters.DungeonCharacter;
 
 public class DaggerStorm implements SpecialAttack {
 
     private int count;
 
     @Override
-    public int specialAttack(DungeonCharacter enemy) {        
+    public int specialAttack(DungeonCharacter enemy) {
     	int dam = 0;
     	for (int i = 0; i < 4; i++) {
             if (Dice.d20() - enemy.getAC() > 0) {
